refactor(tts/magpie/coreml): rename per_module/ -> nanocodec_experiments/
The folder name 'per_module' described the Phase A methodology
(per-module ANE coverage probe) but not the subject. All scripts and
results in this tree are nanocodec-specific (Snake parity, chunked
parity, subgraph probe, etc.). Rename to make the scope explicit.

Updated all in-tree path references and Python imports.
